migrate tests to apidash_core
