+++ packages/apidash_core/test/extensions/string_extensions_test.dart
